ios.maestro: modularize SDK internals
Split the Maestro integration plumbing into focused modules so the public entrypoint reads as orchestration while keeping small helpers close to their owning code.
